Respond to PR feedback
