+++ HelloEFCore/HelloDbContext.cs
@@ -1,7 +1,7 @@
-﻿using Microsoft.EntityFrameworkCore;
-using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel.DataAnnotations;
+using Microsoft.EntityFrameworkCore;
 
-namespace HelloEf
+namespace HelloEFCore
 {
     public class HelloDbContext : DbContext
     {
