+++ HelloEFCore/HelloEFCore.xproj
@@ -7,7 +7,7 @@
   <Import Project="$(VSToolsPath)\DotNet\Microsoft.DotNet.Props" Condition="'$(VSToolsPath)' != ''" />
   <PropertyGroup Label="Globals">
     <ProjectGuid>b9c0f9cf-0d4b-421a-bfca-323f4cb0df0a</ProjectGuid>
-    <RootNamespace>HelloEf</RootNamespace>
+    <RootNamespace>HelloEFCore</RootNamespace>
     <BaseIntermediateOutputPath Condition="'$(BaseIntermediateOutputPath)'=='' ">.\obj</BaseIntermediateOutputPath>
     <OutputPath Condition="'$(OutputPath)'=='' ">.\bin\</OutputPath>
   </PropertyGroup>
